Enable CORS, update GraphQL controller and adjust routes
Enabled CORS by uncommenting the 'rack-cors' gem in the Gemfile, and creating a new cors.rb initializer to handle cross-origin AJAX requests. Updated the GraphQL controller to leverage TaskProAPISchema, enhancing query execution with additional context information, such as the current session. Also adjusted the GraphiQL route in the routes file to support the updated GraphQL controller. These changes help to improve the functionality and usability of the API.