Change team ownership and add user notifications
Made changes to depict more accurately the concept of team ownership in the application by replacing the user association with an owner association in the Teams model. This change provides more clarity on how teams are structured and owned.

Furthermore, a new Notifications model has been created to allow users to receive notifications. This model includes references to the User model, a text field for the notification content, and a boolean to check if the notification has been read.

Updated corresponding database migrations, schema and test files to reflect these changes.