Update README with detailed project information
The README.md file has been significantly updated to provide comprehensive information on the project. Besides the basic details like project name and used technologies, key features and future development tasks have been detailed along with a section illustrating how to get started with the project. This update doesn't change the functionality, it is aimed to give a clear understanding of the project.