Add user authentication with JWT
Added mutations for user sign-up, sign-in, and update along with their respective resolvers in the Users module. Also, added 'generate_jwt' method in the User model to generate JWT for authentication. Modified the graphql controller to handle user authentication logic. Lastly, 'JWT' gem was added for handling JWT and its secret key was added to the credentials.