Added Active Storage for project images, removed teams
We've decided to remove the 'teams' feature due to a shift in our product definition. All associated files and dependencies, including teams and team_memberships tables and models, test files, and references in other model files have been deleted.

On the other hand, we've introduced Active Storage to allow for the attachment of icons to our project records. A new migration for the Active Storage tables has been created, and Active Storage associations and methods have been added to the project model and type. We also updated the .env-template to contain a 'HOST' for constructing image URLs for the icons.

The project update and delete mutations were also updated to reflect these changes.