Refactor GraphQL schema to use Project field
Replaced fields "boards" and "rosters" in GraphQL schema with "project". The reasoning behind this refactor is to simplify the structure, and focus primarily on a projects-oriented schema. Instead of having to deal with multiple endpoints for various entities such as "boards" and "rosters", all related data can now be retrieved through the "project" field by providing the project's ID as an argument. Moreover, this change increases security as it adds a validation to ensure a user is authenticated before granting access to the queried project.