Make password mandatory in UpdateUser mutation
The update functionality for users had been refactored to make the password and confirmation password fields required, increasing the security of updating user details. It simplifies the update process by ensuring the presence of password and confirmation during the update action in the UpdateUser mutation of the user's GraphQL section.